+++ src/Data/CERES/Type.hs
@@ -2,8 +2,10 @@ module Data.CERES.Type where
 
 import qualified Data.Text                     as T
 import           Data.Text.Lazy                 ( Text )
+import           Data.Vector                    ( Vector )
 
 import           TextShow
+import           TextShow.Data.Vector
 
 
 type Name = Text
@@ -13,6 +15,7 @@ type Time = Int
 type ID = Int
 type Idx = Int
 type Message = Text
+type Array a = Vector a
 
 type Branch = T.Text
 type Priority = Int